+++ Model/DashboardRow/PhpSettings.php
@@ -28,50 +28,38 @@ class PhpSettings extends \MageHost\PerformanceDashboard\Model\DashboardRow impl
      */
     public function load()
     {
-        $this->setStatus(self::STATUS_OK);
         $this->setTitle(__('PHP Settings'));
-        $info = '';
-        $problems = '';
-        $warnings = '';
-        $actions = '';
+
         $phpVersionSplit = explode('-', PHP_VERSION, 2);
         $showVersion = reset($phpVersionSplit);
         if (version_compare(PHP_VERSION, '7.0.0', '>=')) {
-            $info .= sprintf(__("PHP Version: %s\n"), $showVersion);
+            $this->info .= sprintf(__("PHP Version: %s\n"), $showVersion);
         } else {
-            $problems .= sprintf(__("PHP Version %s is older than 7.0")."\n", $showVersion);
-            $actions .= __("Upgrade to PHP 7.0 or higher")."\n";
+            $this->problems .= sprintf(__("PHP Version %s is older than 7.0")."\n", $showVersion);
+            $this->actions .= __("Upgrade to PHP 7.0 or higher")."\n";
         }
         foreach ($this->exactValues as $key => $value) {
             $curValue = ini_get($key);
             if (false === $curValue) {
-                $problems .= $this->getProblem($key, $curValue);
-                $actions .= sprintf(__("Set PHP ini setting '%s' to '%s'")."\n", $key, $value);
+                $this->problems .= $this->getProblem($key, $curValue);
+                $this->actions .= sprintf(__("Set PHP ini setting '%s' to '%s'")."\n", $key, $value);
             } elseif (ini_get($key) != $value) {
-                $problems .= $this->getProblem($key, $curValue);
-                $actions .= sprintf(__("Change '%s' to '%s'")."\n", $key, $value);
+                $this->problems .= $this->getProblem($key, $curValue);
+                $this->actions .= sprintf(__("Change '%s' to '%s'")."\n", $key, $value);
             }
         }
         foreach ($this->minimalValues as $key => $value) {
             $curValue = ini_get($key);
             if (false === $curValue) {
-                $problems .= $this->getProblem($key, $curValue);
-                $actions .= sprintf(__("Set PHP setting '%s' to '%s' or higher")."\n", $key, $value);
+                $this->problems .= $this->getProblem($key, $curValue);
+                $this->actions .= sprintf(__("Set PHP setting '%s' to '%s' or higher")."\n", $key, $value);
             } elseif (ini_get($key) < $value) {
-                $problems .= $this->getProblem($key, $curValue);
-                $actions .= sprintf(__("Change '%s' to '%s' or higher")."\n", $key, $value);
+                $this->problems .= $this->getProblem($key, $curValue);
+                $this->actions .= sprintf(__("Change '%s' to '%s' or higher")."\n", $key, $value);
             }
         }
 
-        if ($problems) {
-            $this->setStatus(self::STATUS_PROBLEM);
-        } elseif ($warnings) {
-            $this->setStatus(self::STATUS_WARNING);
-        } else {
-            $this->setStatus(self::STATUS_OK);
-        };
-        $this->setInfo($problems.$warnings.$info);
-        $this->setAction($actions);
+        $this->groupProcess();
     }
 
     private function getProblem($key, $curValue)